+++ pwn/toplevel.py
@@ -73,8 +73,8 @@ from pwnlib.util.web import *
 
 # Promote these modules, so that "from pwn import *" will let you access them
 
-from six.moves import cPickle as pickle, cStringIO as StringIO
-from six import BytesIO
+import pickle
+from io import BytesIO, StringIO
 
 log = getLogger("pwnlib.exploit")
 error   = log.error